Add utility method to automate BigInt/Buffer conversion in objects
