+++ services/core/shared/core/compat/common/index.js
@@ -16,6 +16,10 @@
 const http = require('./httpRequest');
 
 const {
+    getApiClient,
+} = require('./wsRequest');
+
+const {
     getNetworkConstants,
     setCoreVersion,
     getCoreVersion,
@@ -33,8 +37,14 @@ const {
     validateTimestamp,
 } = require('./time');
 
+const {
+    parseToJSONCompatObj,
+} = require('./utils');
+
 module.exports = {
     http,
+
+    getApiClient,
 
     getNetworkConstants,
     setCoreVersion,
@@ -49,4 +59,6 @@ module.exports = {
     getEpochUnixTime,
     getUnixTime,
     validateTimestamp,
+
+    parseToJSONCompatObj,
 };
