+++ services/core/shared/core/compat/sdk_v5/accounts.js
@@ -14,10 +14,10 @@
  *
  */
 const BluebirdPromise = require('bluebird');
+
 const coreApi = require('./coreApi');
-
 const coreCache = require('./coreCache');
-
+const { parseToJSONCompatObj } = require('../common');
 // const { getDelegates } = require('./delegates');
 
 // const balanceUnlockWaitHeightSelf = 260000;
@@ -68,17 +68,6 @@ const confirmSecondPublicKey = async secondPublicKey => {
 // 	return accounts;
 // };
 
-const parseBigIntToNumber = obj => {
-	const result = {};
-	Object.entries(obj)
-		.forEach(([k, v]) => {
-			if (typeof v === 'object' && v !== null) result[k] = parseBigIntToNumber(v);
-			else if (typeof v === 'bigint') result[k] = Number(v);
-			else result[k] = v;
-		});
-	return result;
-};
-
 const normalizeAccount = account => {
 	account.address = account.address.toString('hex');
 	account.isDelegate = !!(account.dpos && account.dpos.delegate);
@@ -93,7 +82,7 @@ const normalizeAccount = account => {
 			return vote;
 		});
 
-	return parseBigIntToNumber(account);
+	return parseToJSONCompatObj(account);
 };
 
 const getAccounts = async params => {
